+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - If your ChurchSuite account contains a plan item category named "hidden", hide items
   of that category
+- Links within layouts will open in the default browser
 
 ### Changed